requirements.txt
aiohttp==3.9.5
aiosignal==1.3.1
astunparse==1.6.3
attrs==23.2.0
braceexpand==0.1.7
cachetools==5.3.3
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
discord==2.3.2
discord-ext-bot==1.0.1
discord.py==2.3.2
frozenlist==1.4.1
google-api-core==2.18.0
google-api-python-client==2.127.0
google-auth==2.29.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.0
googleapis-common-protos==1.63.0
gspread==6.1.0
httplib2==0.22.0
idna==3.7
import-expression==1.1.4
install==1.3.5
jishaku==2.5.2
multidict==6.0.5
oauthlib==3.2.2
proto-plus==1.23.0
protobuf==4.25.3
py-cord==2.5.0
pyasn1==0.6.0
pyasn1_modules==0.4.0
pyparsing==3.1.2
python-dotenv==1.0.1
requests==2.31.0
requests-oauthlib==2.0.0
rsa==4.9
six==1.16.0
StrEnum==0.4.15
uritemplate==4.1.1
urllib3==2.2.1
yarl==1.9.4
